+++ package.json
@@ -10,7 +10,8 @@
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "3.4.1"
+    "react-scripts": "3.4.1",
+    "sass-node": "^4.11.2"
   },
   "devDependencies": {
     "@mate-academy/eslint-config-react": "*",
